pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pm_config_loader"
version = "0.0.6"
description = "This is a simple ConfigLoader with possibility to do a Env replace."
authors = [
  { name="Patrick Maul", email="maulpatrick1996@gmail.com" },
]
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "PyYAML==3.11",
]

[project.urls]
"Homepage" = "https://github.com/PatrickMaul/config_loader_pkg"
"Bug Tracker" = "https://github.com/PatrickMaul/config_loader_pkg/issues"

[tool.poetry]
include = [
  "src"
]
exclude = [
  "test",
  "assets"
]